requirements.txt
google-api-python-client
pandas
apache-airflow
snowflake-connector-python
snowflake-sqlalchemy
apache-airflow-providers-snowflake